% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonNegativeMatrixFactorization.R
\name{ColumnCentralizeSparseMatrix}
\alias{ColumnCentralizeSparseMatrix}
\title{Sparse matrix column centralization.}
\usage{
ColumnCentralizeSparseMatrix(smat, centerFinder = median, spreadFinder = NULL)
}
\arguments{
\item{smat}{a sparse item-tag (document-term) matrix}

\item{centerFinder}{a function used to find the center of a numerical vector}

\item{spreadFinder}{a function to find the spread of each vector and divide its the centered elements}
}
\description{
Centralizes a sparse matrix around the mean of the non-zero elements in each column.
The sparse matrix zero entries are also zeroes in the centralized matrix.
}
\details{
If the argument spreadFinder is NULL no scaling of the centered elements is done.
It would be nice the argument \code{spreadFinder} can take as argument also the strings
"interQuartile" and "standardDeviation".
}
