% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonNegativeMatrixFactorization.R
\name{NearestWords}
\alias{NearestWords}
\title{Statistical thesaurus entry calculation.}
\usage{
NearestWords(H, word, n = 20, fixed = TRUE)
}
\arguments{
\item{H}{A matrix of topics (each row is a topic, each column is a word).}

\item{word}{An integer index or a pattern for a word (a column of H).}

\item{n}{Number of nearest neighbors to be found (size of the thesaurus entry).}

\item{fixed}{If TRUE word is taken as a fixed pattern; if FALSE as a regexp.}
}
\value{
Returns a data frame with columns c("Distance", "Index", "Word").
}
\description{
Statistical thesaurus entry calculation for a specified matrix of topics and a word.
}
\details{
Euclidean distance is used to find the nearest neighbors.
}
