% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonNegativeMatrixFactorization.R
\name{NNMF}
\alias{NNMF}
\title{Non-Negative Matrix Factorization (NNMF).}
\usage{
NNMF(V, k, maxSteps = 200, nonNegativeQ = TRUE, epsilon = 10^-9,
  regularizationParameter = 0.01, tolerance = 1e-04,
  profilingQ = FALSE)
}
\arguments{
\item{V}{A sparse matrix.}

\item{k}{Rank for the factorization.}

\item{maxSteps}{Maximum steps for the iteration process}

\item{nonNegativeQ}{Should the factorization be non-negative or not?}

\item{epsilon}{Denominator offset.}

\item{regularizationParameter}{Regularization parameter (diagonal matrix product).}

\item{tolerance}{Tolerance for the factorization approximation.}

\item{profilingQ}{Should profiling information be printed during execution or not?}
}
\value{
A list with named elements corresponding to the matrix factors.
}
\description{
Returns the pair of matrices {W,H} such that V = W H and
the number of the columns of W and the number of rows of H are k.
The method used is Gradient Descent with Constrained Least Squares.
}
