% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonNegativeMatrixFactorization.R
\name{NNMFBasisVectorInterpretation}
\alias{NNMFBasisVectorInterpretation}
\title{Basis vector interpretation.}
\usage{
NNMFBasisVectorInterpretation(vec, n, interpretationTerms = NULL)
}
\arguments{
\item{n}{Number of elements to be interpreted.}

\item{interpretationTerms}{A character vector of interpretation terms.}
}
\value{
A numeric vector with named elements.
}
\description{
Takes the \code{n} largest coordinates of \code{vec},
finds the corresponding elements in a list of interpretation terms,
corresponding to the vector coordinates.
By default the interpretation terms are the names of the vector elements.
}
