% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonNegativeMatrixFactorization.R
\name{NNMFNormalizeMatrixProduct}
\alias{NNMFNormalizeMatrixProduct}
\title{Normalization of a matrix product.}
\usage{
NNMFNormalizeMatrixProduct(W, H, normalizeLeftQ = TRUE)
}
\arguments{
\item{W}{The right matrix factor.}

\item{H}{The left matrix factor.}

\item{normalizeLeftQ}{Is the left factor normalized?}
}
\value{
A list with named elements corresponding to the matrix factors.
}
\description{
Normalization of a matrix product.
}
