% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonNegativeMatrixFactorization.R
\name{RowCentralizeSparseMatrix}
\alias{RowCentralizeSparseMatrix}
\title{Sparse matrix row centralization.}
\usage{
RowCentralizeSparseMatrix(smat, centerFinder = median, spreadFinder = NULL)
}
\arguments{
\item{smat}{A sparse item-tag (document-term) matrix.}

\item{centerFinder}{A function used to find the center of a numerical vector.}

\item{spreadFinder}{A function to find the spread of each vector and divide its the centered elements.}
}
\value{
A sparse matrix.
}
\description{
Cenralizes a sparse matrix around the mean of the non-zero elements in each row
The zero entries are also zeroes in the centralized matrix.
}
