% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonNegativeMatrixFactorization.R
\name{SparseMatrixExponent}
\alias{SparseMatrixExponent}
\title{Sparse matrix exponent}
\usage{
SparseMatrixExponent(smat, exp = -1, tol = 10^-6, ...)
}
\arguments{
\item{smat}{Sparse matrix.}

\item{exp}{Exponent (a real number.)}

\item{tol}{Tolerance (a real number.)}

\item{...}{Additional arguments for \code{\link{svd}}.}
}
\value{
Sparse matrix
}
\description{
Computes the pseudo inverse of a sparse matrix.
}
\details{
Uses SVD through \code{\link{svd}}.
}
