% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonNegativeMatrixFactorization.R
\name{SparseMatrixQ}
\alias{SparseMatrixQ}
\title{Sparse matrix test.}
\usage{
SparseMatrixQ(object)
}
\arguments{
\item{object}{An object.}
}
\value{
A logical value.
}
\description{
Tests is the argument one of sparse matrix classes in
the base package \link{\code{Matrix}}.
}
\details{
These class names are considered sparse matrix classes:
\code{names(getClass("Matrix")@subclasses)}.
}
